+++ src/EventListener/SearchIndexerSubscriber.php
@@ -25,40 +25,16 @@ class SearchIndexerSubscriber implements EventSubscriber
 
     public function postUpdate(LifecycleEventArgs $args)
     {
-        $object = $args->getObject();
-        $objectManager = $args->getObjectManager();
-
-        if (!$this->indexManager->isSearchable($object)) {
-            return;
-        }
-
-        if ($this->indexManager->shouldBeIndexed($object)) {
-            $this->indexManager->index($object, $objectManager);
-        } else {
-            $this->indexManager->remove($object, $objectManager);
-        }
+        $this->indexManager->index($args->getObject(), $args->getObjectManager());
     }
 
     public function postPersist(LifecycleEventArgs $args)
     {
-        $object = $args->getObject();
-        $objectManager = $args->getObjectManager();
-
-        if (!$this->indexManager->isSearchable($object)) {
-            return;
-        }
-
-        if ($this->indexManager->shouldBeIndexed($object)) {
-            $this->indexManager->index($object, $objectManager);
-        }
+        $this->indexManager->index($args->getObject(), $args->getObjectManager());
     }
 
     public function preRemove(LifecycleEventArgs $args)
     {
-        $object = $args->getObject();
-
-        if ($this->indexManager->isSearchable($object)) {
-            $this->indexManager->remove($object, $args->getObjectManager());
-        }
+        $this->indexManager->remove($object = $args->getObject(), $args->getObjectManager());
     }
 }