+++ src/Exception/InvalidEntityForAggregator.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Algolia\SearchBundle\Exception;
+
+use LogicException;
+
+class InvalidEntityForAggregator extends LogicException
+{
+}